Move package client.okhttp to okhttp.client
